requirements.txt
absl-py==0.12.0
asgiref==3.3.4
astunparse==1.6.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Django==3.2.2
flatbuffers==1.12
future==0.18.2
gast==0.4.0
geographiclib==1.50
geopy==2.1.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
image==1.5.33
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.2.52
opt-einsum==3.3.0
pandas==1.2.4
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s2sphere==0.2.5
six==1.15.0
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc3
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
tf==1.0.0
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1

folium~=0.12.1
tensorflow-gpu~=2.5.0
ipython~=7.25.0
goslate~=1.5.1
PyDictionary~=2.0.1
flickrapi~=2.4.0
foursquare~=1!2020.1.30
googletrans~=2.3.0
setuptools~=56.1.0
scikit-learn~=0.24.2